extrait le bouton "..." et son menu déroulant pour ré-utilisation

+++ app.territoiresentransitions.react/src/app/pages/collectivite/Referentiels/ActionDiscussions/ActionDiscussionCommentaireDropdown.tsx
@@ -1,65 +1,34 @@
-import {forwardRef, Ref} from 'react';
-import IconThreeDotHorizontal from 'ui/icons/IconThreeDotHorizontal';
-
-import DropdownFloater from 'ui/shared/floating-ui/DropdownFloater';
+import ThreeDotMenu from 'ui/shared/select/ThreeDotMenu';
 import {useDeleteCommentaireFromDiscussion} from './data/useDeleteCommentaireFromDiscussion';
 
 type Props = {
   commentaire_id: number;
 };
 
+const OPTIONS = [
+  {
+    value: 'delete',
+    label: 'Supprimer mon commentaire',
+    icon: 'fr-icon-delete-line',
+  },
+];
+
 /** Menu et options pour un commentaire dans une discussion */
 const ActionDiscussionCommentaireDropdown = ({commentaire_id}: Props) => {
   const {mutate: deleteCommentaire} = useDeleteCommentaireFromDiscussion();
 
   return (
-    <>
-      <DropdownFloater
-        placement="bottom-end"
-        render={({close}) => (
-          <nav data-test="ActionDiscussionCommentaireMenu">
-            <ul className="m-0 p-0">
-              <li className="fr-nav__item pb-0 border-b border-gray-200">
-                <button
-                  className="fr-nav__link !flex !items-center !text-bf500 !py-2 !px-3 before:!hidden !shadow-none"
-                  onClick={() => {
-                    deleteCommentaire(commentaire_id);
-                    close();
-                  }}
-                >
-                  <span className="fr-fi-delete-line scale-75 mr-1" />
-                  <span>Supprimer mon commentaire</span>
-                </button>
-              </li>
-            </ul>
-          </nav>
-        )}
-      >
-        <ActionDiscussionCommentaireDropdownButtonDisplayed />
-      </DropdownFloater>
-    </>
+    <ThreeDotMenu
+      options={OPTIONS}
+      onSelect={value => {
+        if (value === 'delete') {
+          deleteCommentaire(commentaire_id);
+        }
+      }}
+      dataTest="ActionDiscussionCommentaireMenu"
+      dataTestButton="ActionDiscussionCommentaireMenuButton"
+    />
   );
 };
 
 export default ActionDiscussionCommentaireDropdown;
-
-const ActionDiscussionCommentaireDropdownButtonDisplayed = forwardRef(
-  (
-    {
-      isOpen,
-      ...props
-    }: {
-      isOpen?: boolean;
-    },
-    ref?: Ref<HTMLDivElement>
-  ) => (
-    <div ref={ref} className="shrink-0 flex border border-gray-200" {...props}>
-      <button
-        data-test="ActionDiscussionCommentaireMenuButton"
-        className="p-1 text-bf500"
-      >
-        <IconThreeDotHorizontal className="w-5 h-5 fill-bf500" />
-      </button>
-    </div>
-  )
-);
